// src/API/NauthApi_gen/models/EmailTemplateDTO.ts
/* tslint:disable */
/* eslint-disable */
/**
 * nauth-asp
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from "../runtime";
import type { EmailTemplateType } from "./EmailTemplateType";
import {
  EmailTemplateTypeFromJSON,
  EmailTemplateTypeFromJSONTyped,
  EmailTemplateTypeToJSON,
  EmailTemplateTypeToJSONTyped,
} from "./EmailTemplateType";

/**
 *
 * @export
 * @interface EmailTemplateDTO
 */
export interface EmailTemplateDTO {
  /**
   *
   * @type {string}
   * @memberof EmailTemplateDTO
   */
  id?: string | null;
  /**
   *
   * @type {string}
   * @memberof EmailTemplateDTO
   */
  name?: string | null;
  /**
   *
   * @type {boolean}
   * @memberof EmailTemplateDTO
   */
  isActive?: boolean;
  /**
   *
   * @type {EmailTemplateType}
   * @memberof EmailTemplateDTO
   */
  type?: EmailTemplateType;
  /**
   *
   * @type {string}
   * @memberof EmailTemplateDTO
   */
  subject?: string | null;
  /**
   *
   * @type {string}
   * @memberof EmailTemplateDTO
   */
  body?: string | null;
  /**
   *
   * @type {string}
   * @memberof EmailTemplateDTO
   */
  htmlBody?: string | null;
  /**
   *
   * @type {Date}
   * @memberof EmailTemplateDTO
   */
  createdAt?: Date;
}

/**
 * Check if a given object implements the EmailTemplateDTO interface.
 */
export function instanceOfEmailTemplateDTO(value: object): value is EmailTemplateDTO {
  return true;
}

export function EmailTemplateDTOFromJSON(json: any): EmailTemplateDTO {
  return EmailTemplateDTOFromJSONTyped(json, false);
}

export function EmailTemplateDTOFromJSONTyped(json: any, ignoreDiscriminator: boolean): EmailTemplateDTO {
  if (json == null) {
    return json;
  }
  return {
    id: json["id"] == null ? undefined : json["id"],
    name: json["name"] == null ? undefined : json["name"],
    isActive: json["isActive"] == null ? undefined : json["isActive"],
    type: json["type"] == null ? undefined : EmailTemplateTypeFromJSON(json["type"]),
    subject: json["subject"] == null ? undefined : json["subject"],
    body: json["body"] == null ? undefined : json["body"],
    htmlBody: json["htmlBody"] == null ? undefined : json["htmlBody"],
    createdAt: json["createdAt"] == null ? undefined : new Date(json["createdAt"]),
  };
}

export function EmailTemplateDTOToJSON(json: any): EmailTemplateDTO {
  return EmailTemplateDTOToJSONTyped(json, false);
}

export function EmailTemplateDTOToJSONTyped(value?: EmailTemplateDTO | null, ignoreDiscriminator: boolean = false): any {
  if (value == null) {
    return value;
  }

  return {
    id: value["id"],
    name: value["name"],
    isActive: value["isActive"],
    type: EmailTemplateTypeToJSON(value["type"]),
    subject: value["subject"],
    body: value["body"],
    htmlBody: value["htmlBody"],
    createdAt: value["createdAt"] == null ? undefined : value["createdAt"].toISOString(),
  };
}