// src/API/NauthApi_gen/models/EmailTemplateType.ts
/* tslint:disable */
/* eslint-disable */
/**
 * nauth-asp
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 *
 * @export
 */
export const EmailTemplateType = {
  VerifyEmail: "VerifyEmail",
  PasswordReset: "PasswordReset",
  EmailCode: "EmailCode",
  EmailSignIn: "EmailSignIn",
  DeleteAccount: "DeleteAccount",
  ChangeEmail: "ChangeEmail",
} as const;
export type EmailTemplateType = (typeof EmailTemplateType)[keyof typeof EmailTemplateType];

export function instanceOfEmailTemplateType(value: any): boolean {
  for (const key in EmailTemplateType) {
    if (Object.prototype.hasOwnProperty.call(EmailTemplateType, key)) {
      if (EmailTemplateType[key as keyof typeof EmailTemplateType] === value) {
        return true;
      }
    }
  }
  return false;
}

export function EmailTemplateTypeFromJSON(json: any): EmailTemplateType {
  return EmailTemplateTypeFromJSONTyped(json, false);
}

export function EmailTemplateTypeFromJSONTyped(json: any, ignoreDiscriminator: boolean): EmailTemplateType {
  return json as EmailTemplateType;
}

export function EmailTemplateTypeToJSON(value?: EmailTemplateType | null): any {
  return value as any;
}

export function EmailTemplateTypeToJSONTyped(value: any, ignoreDiscriminator: boolean): EmailTemplateType {
  return value as EmailTemplateType;
}